Add mouseover display of read name to pileup
